libobs: Fix sign-compare warning
`diff` is used only if greater than `interval_ns` which is unsigned.
So casting it as unsigned will not change its sign.

+++ libobs/obs-video.c
@@ -797,8 +797,9 @@ static inline void video_sleep(struct obs_core_video *video, bool raw_active,
 		const uint64_t udiff = os_gettime_ns() - cur_time;
 		int64_t diff;
 		memcpy(&diff, &udiff, sizeof(diff));
-		const uint64_t clamped_diff =
-			(diff > (int64_t)interval_ns) ? diff : interval_ns;
+		const uint64_t clamped_diff = (diff > (int64_t)interval_ns)
+						      ? (uint64_t)diff
+						      : interval_ns;
 		count = (int)(clamped_diff / interval_ns);
 		*p_time = cur_time + interval_ns * count;
 	}
